# Optimised the libSession size limit tests