- Optimised the libSession size limit tests
- Tweaked another couple failing tests

Some quick stats on the differences with the updated tests using an M1 Pro:
|   | Old | New | Reduction |
| ------------- | ------------- | ------------- | ------------- |
| **LibSessionUtilSpec libSession, CONTACTS, when checking error catching, can catch size limit errors thrown when pushing.** | 9.015 seconds | 1.877 seconds | ~80% |
| **LibSessionUtilSpec libSession, CONTACTS, when checking size limits, has not changed the max filled records.** | 3.781 seconds | 1.807 seconds | ~52% |
| **LibSessionUtilSpec libSession, CONTACTS, when checking size limits, has not changed the max name and profile pic only records.** | 3.949 seconds | 1.554 seconds | ~60% |
| **LibSessionUtilSpec libSession, CONTACTS, when checking size limits, has not changed the max name only records.**  | 13.704 seconds  | 1.666 seconds | ~88% |
| **LibSessionUtilSpec libSession, CONTACTS, when checking size limits, has not changed the max empty records.** | 44.096 seconds | 3.549 seconds | ~92% |